Add Shepherd Embryo v0.0002 - topology-first architecture
